requirements.txt
aiopg==1.4.0 ; python_version >= '3.7'
asgiref==3.8.1 ; python_version >= '3.8'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==24.2.0 ; python_version >= '3.7'
autobahn==24.4.2 ; python_version >= '3.9'
automat==24.8.1 ; python_version >= '3.8'
certifi==2024.8.30 ; python_version >= '3.6'
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
channels==4.0.0
channels-postgres==1.0.5
charset-normalizer==3.4.0 ; python_full_version >= '3.7.0'
constantly==23.10.4 ; python_version >= '3.8'
cryptography==44.0.0 ; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.1.2
django==5.1.3
django-cors-headers==4.6.0
django-sass==1.1.0
hyperlink==21.0.0
idna==3.10 ; python_version >= '3.6'
incremental==24.7.2 ; python_version >= '3.8'
libsass==0.23.0 ; python_version >= '3.8'
msgpack==1.0.8 ; python_version >= '3.8'
oauthlib==3.2.2 ; python_version >= '3.6'
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10 ; python_version >= '3.8'
pyasn1==0.6.1 ; python_version >= '3.8'
pyasn1-modules==0.4.1 ; python_version >= '3.8'
pycparser==2.22 ; python_version >= '3.8'
pyopenssl==24.3.0
python-dotenv==1.0.1
reloadium==1.5.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0 ; python_version >= '3.9'
sqlparse==0.5.2 ; python_version >= '3.8'
twisted[tls]==24.10.0 ; python_full_version >= '3.8.0'
txaio==23.1.1 ; python_version >= '3.7'
typing-extensions==4.12.2 ; python_version >= '3.8'
urllib3==2.2.3 ; python_version >= '3.8'
zope-interface==7.2 ; python_version >= '3.8'
django-otp==1.5.4
qrcode==8.0
pyotp==2.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0